ref(perf-view): Convert transaction summary header into component
Extract the header of the transaction summary page into its own component.
This is in preparation for the upcoming rum page.
