+++ src/sentry/static/sentry/app/views/performance/transactionSummary/header.tsx
@@ -1,0 +1,103 @@
+import React from 'react';
+import {Location} from 'history';
+
+import {Organization, Project} from 'app/types';
+import EventView from 'app/utils/discover/eventView';
+import Feature from 'app/components/acl/feature';
+import CreateAlertButton from 'app/components/createAlertButton';
+import * as Layout from 'app/components/layouts/thirds';
+import ButtonBar from 'app/components/buttonBar';
+import {trackAnalyticsEvent} from 'app/utils/analytics';
+import Breadcrumb from 'app/views/performance/breadcrumb';
+
+import KeyTransactionButton from './keyTransactionButton';
+
+type Props = {
+  eventView: EventView;
+  location: Location;
+  organization: Organization;
+  projects: Project[];
+  transactionName: string;
+};
+
+class TransactionHeader extends React.Component<Props> {
+  trackAlertClick(errors?: Record<string, boolean>) {
+    const {organization} = this.props;
+    trackAnalyticsEvent({
+      eventKey: 'performance_views.summary.create_alert_clicked',
+      eventName: 'Performance Views: Create alert clicked',
+      organization_id: organization.id,
+      status: errors ? 'error' : 'success',
+      errors,
+      url: window.location.href,
+    });
+  }
+
+  handleIncompatibleQuery: React.ComponentProps<
+    typeof CreateAlertButton
+  >['onIncompatibleQuery'] = (incompatibleAlertNoticeFn, errors) => {
+    this.trackAlertClick(errors);
+    const incompatibleAlertNotice = incompatibleAlertNoticeFn(() =>
+      this.setState({incompatibleAlertNotice: null})
+    );
+    this.setState({incompatibleAlertNotice});
+  };
+
+  handleCreateAlertSuccess = () => {
+    this.trackAlertClick();
+  };
+
+  renderCreateAlertButton() {
+    const {eventView, organization, projects} = this.props;
+
+    return (
+      <CreateAlertButton
+        eventView={eventView}
+        organization={organization}
+        projects={projects}
+        onIncompatibleQuery={this.handleIncompatibleQuery}
+        onSuccess={this.handleCreateAlertSuccess}
+        referrer="performance"
+      />
+    );
+  }
+
+  renderKeyTransactionButton() {
+    const {eventView, organization, transactionName} = this.props;
+
+    return (
+      <KeyTransactionButton
+        transactionName={transactionName}
+        eventView={eventView}
+        organization={organization}
+      />
+    );
+  }
+
+  render() {
+    const {organization, location, transactionName} = this.props;
+
+    return (
+      <Layout.Header>
+        <Layout.HeaderContent>
+          <Breadcrumb
+            organization={organization}
+            location={location}
+            transactionName={transactionName}
+          />
+          <Layout.Title>{transactionName}</Layout.Title>
+        </Layout.HeaderContent>
+        <Layout.HeaderActions>
+          <ButtonBar gap={1}>
+            <Feature organization={organization} features={['incidents']}>
+              {({hasFeature}) => hasFeature && this.renderCreateAlertButton()}
+            </Feature>
+            {this.renderKeyTransactionButton()}
+          </ButtonBar>
+        </Layout.HeaderActions>
+      </Layout.Header>
+    );
+  }
+}
+
+export default TransactionHeader;
